Add single-gallery view
Using plugins on grid view is cumbersome, user needs to scroll the grid
so much, because Vaadin grid does not allow to scroll to a selected
(next) row. This was the reason to create a single-gallery view, which
is totally under developer's control, allows automatic switching to next
gallery after plugin action (not implemented yet) which improves
usability of certain use cases (e.g. removing or rating).
Currently both - grid view and single view - are supported.
Eventually grid view should be used for general viewing and single view
for editing.